DucklingManager
Implemented the addDuck() function. Currently, the duckling is just a triangle sprite for testing. The code counts how many fish were eaten, adds a duckling object for every 2 fish, and makes sure each duck spawn behind all other ducks while maintaining distance and speed, both variables are adjustable through unity.

Also, added an if statement to check for duck count if it reaches fishPerLevel. If so, it should call a function that is yet to be implemented, nextLevel();

TO-DO:
-Add a duckling sprite--instead of the triangle--to the duckling game object.
-Adjust size, speed, and distance of the duckling.
-Implement NextLevel();